fix: move ionosclients to clients
