+++ internal/clients/ionoscloud.go
@@ -1,4 +1,4 @@
-package ionosclients
+package clients
 
 import (
 	"context"
